+++ src/core/types/pre_checkout_query.rs
@@ -5,7 +5,7 @@ pub struct PreCheckoutQuery {
     pub id: String,
     pub from: User,
     pub currency: String,
-    pub total_amount: i64,
+    pub total_amount: i32,
     pub invoice_payload: String,
     pub shipping_option_id: Option<String>,
     pub order_info: Option<OrderInfo>,
